apps/vglnt-data/pyproject.toml
[project]
name = "vglnt-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "opencv-python>=4.10.0.84",
    "python-dotenv>=1.0.1",
    "yt-dlp>=2024.12.13",
]
